Add functionality for all nums / basic operations
Implement event listener by selecting the button container, and returning if the button isn't pressed
If statements within the event listener dictate the calculations, for each button type. Any calculations will set the past calculation into pastCalc
A number button click will check if the first number is complete, if not, add that digit to the number. If it is, then add the digit to the second number

An operator button click will check if the left side / first number is true, if not, update it as true, set event.id as the operator, and add it to the display.
If the left side is already true, and a second number exists, it will calculate the first number, operator, and second number, appending this operator to the result for further calculations

An equal button click will calculate the current calculation in the display, returning the result back to the display for further calculation, resetting operator and num2.
If the calculation does not contain a num2, then it will simply return num1 as result, and display it for further calculations.

Oh and I made the display look nicer (color diff between current and pastCalc, different fonts for buttons and display)